Make sure it doesn't crash without license
